move dxl call into a custom api
