+++ alexk_low_cost_robot/nodes/lcr_teleoperate.py
@@ -3,249 +3,11 @@ import os
 import numpy as np
 import pyarrow as pa
 
-from dynamixel_sdk import PacketHandler, PortHandler, COMM_SUCCESS
 from dora import Node
+from dynamixel_sdk import PortHandler, PacketHandler
 
-
-def write_goal_current(packet: PacketHandler, serial: PortHandler, id: int, goal_current):
-    """
-    Write the goal current to the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    :param goal_current: current
-    """
-    comm, error = packet.write2ByteTxRx(serial, id, 102, goal_current)
-
-    if goal_current is not None:
-        if comm != COMM_SUCCESS:
-            print(f"Failed to communicate with motor {id}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to write goal current {goal_current} to motor {id}")
-            print("%s" % packet.getRxPacketError(error))
-
-
-def read_present_current(packet: PacketHandler, serial: PortHandler, id: int):
-    """
-    Read the present current from the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    :return: int
-    """
-
-    current, comm, error = packet.read2ByteTxRx(serial, id, 126)
-
-    if comm != COMM_SUCCESS:
-        print(f"Failed to communicate with motor {id}")
-        print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to read present current from motor {id}")
-        print("%s" % packet.getRxPacketError(error))
-
-    return current if comm == COMM_SUCCESS and error == 0 else None
-
-
-def write_goal_position(packet: PacketHandler, serial: PortHandler, id: int, goal_position):
-    """
-    Write the goal position to the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    :param goal_position:
-    """
-    comm, error = packet.write4ByteTxRx(serial, id, 116, goal_position)
-
-    if goal_position is not None:
-        if comm != COMM_SUCCESS:
-            print(f"Failed to communicate with motor {id}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to write goal position {goal_position} to motor {id}")
-            print("%s" % packet.getRxPacketError(error))
-
-
-def write_goal_positions(packet: PacketHandler, serial: PortHandler, ids: np.array, goal_positions: np.array):
-    """
-    Write the goal positions to the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param ids: np.array
-    :param goal_positions: np.array
-    """
-
-    for i in range(len(ids)):
-        if goal_positions[i] is not None:
-            comm, error = packet.write4ByteTxRx(serial, ids[i], 116, goal_positions[i])
-            if comm != COMM_SUCCESS:
-                print(f"Failed to communicate with motor {ids[i]}")
-                print("%s" % packet.getTxRxResult(comm))
-            if error != 0:
-                print(f"Failed to write goal position {goal_positions[i]} to motor {ids[i]}")
-                print("%s" % packet.getRxPacketError(error))
-
-
-def read_present_position(packet: PacketHandler, serial: PortHandler, id: int):
-    """
-    Read the present position from the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    :return: int
-    """
-
-    position, comm, error = packet.read4ByteTxRx(serial, id, 132)
-
-    if comm != COMM_SUCCESS:
-        print(f"Failed to communicate with motor {id}")
-        print("%s" % packet.getTxRxResult(comm))
-    if error != 0:
-        print(f"Failed to read present position from motor {id}")
-        print("%s" % packet.getRxPacketError(error))
-
-    return position if comm == COMM_SUCCESS and error == 0 else None
-
-
-def read_present_positions(packet: PacketHandler, serial: PortHandler, ids: np.array):
-    """
-    Read the present positions from the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param ids: np.array
-    :return: np.array
-    """
-    present_positions = []
-
-    for id_ in ids:
-        position, comm, error = packet.read4ByteTxRx(serial, id_, 132)
-
-        if comm != COMM_SUCCESS:
-            print(f"Failed to communicate with motor {id_}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to read present position from motor {id_}")
-            print("%s" % packet.getRxPacketError(error))
-
-        present_positions.append(position if comm == COMM_SUCCESS and error == 0 else None)
-
-    return np.array(present_positions)
-
-
-def read_present_velocity(packet: PacketHandler, serial: PortHandler, id: int):
-    """
-    Read the present velocity from the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    :return: int
-    """
-
-    velocity, comm, error = packet.read4ByteTxRx(serial, id, 128)
-
-    if comm != COMM_SUCCESS:
-        print(f"Failed to communicate with motor {id}")
-        print("%s" % packet.getTxRxResult(comm))
-    if error != 0:
-        print(f"Failed to read present velocity from motor {id}")
-        print("%s" % packet.getRxPacketError(error))
-
-    return velocity if comm == COMM_SUCCESS and error == 0 else None
-
-
-def read_present_velocities(packet: PacketHandler, serial: PortHandler, ids: np.array):
-    """
-    Read the present velocities from the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param ids: list
-    :return: list
-    """
-    present_velocities = []
-
-    for id_ in ids:
-        velocity, comm, error = packet.read4ByteTxRx(serial, id_, 128)
-
-        if comm != COMM_SUCCESS:
-            print(f"Failed to communicate with motor {id_}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to read present velocity from motor {id_}")
-            print("%s" % packet.getRxPacketError(error))
-
-        present_velocities.append(velocity if comm == COMM_SUCCESS and error == 0 else None)
-
-    return np.array(present_velocities)
-
-
-def enable_torque(packet: PacketHandler, serial: PortHandler, id: int):
-    """
-    Enable the torque of the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    """
-    comm, error = packet.write1ByteTxRx(serial, id, 64, 1)
-
-    if comm != COMM_SUCCESS:
-        print(f"Failed to communicate with motor {id}")
-        print("%s" % packet.getTxRxResult(comm))
-    if error != 0:
-        print(f"Failed to enable torque for motor {id}")
-        print("%s" % packet.getRxPacketError(error))
-
-
-def enable_torques(packet: PacketHandler, serial: PortHandler, ids: np.array):
-    """
-    Enable the torques of the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param ids: np.array
-    """
-    for id_ in ids:
-        comm, error = packet.write1ByteTxRx(serial, id_, 64, 1)
-
-        if comm != COMM_SUCCESS:
-            print(f"Failed to communicate with motor {id_}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to enable torque for motor {id_}")
-            print("%s" % packet.getRxPacketError(error))
-
-
-def disable_torque(packet: PacketHandler, serial: PortHandler, id: int):
-    """
-    Disable the torque of the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param id: int
-    """
-    comm, error = packet.write1ByteTxRx(serial, id, 64, 0)
-
-    if comm != COMM_SUCCESS:
-        print(f"Failed to disable torque for motor {id}")
-        print("%s" % packet.getTxRxResult(comm))
-    if error != 0:
-        print(f"Failed to disable torque for motor {id}")
-        print("%s" % packet.getRxPacketError(error))
-
-
-def disable_torques(packet: PacketHandler, serial: PortHandler, ids: np.array):
-    """
-    Disable the torques of the puppet robot
-    :param packet: PacketHandler
-    :param serial: PortHandler
-    :param ids: np.array
-    """
-    for id_ in ids:
-        comm, error = packet.write1ByteTxRx(serial, id_, 64, 0)
-
-        if comm != COMM_SUCCESS:
-            print(f"Failed to disable torque for motor {id_}")
-            print("%s" % packet.getTxRxResult(comm))
-        if error != 0:
-            print(f"Failed to disable torque for motor {id_}")
-            print("%s" % packet.getRxPacketError(error))
+from alexk_arm import write_goal_current, read_present_current, write_goal_position, read_present_positions, \
+    enable_torque, enable_torques, write_goal_positions, read_present_velocities, disable_torque, disable_torques
 
 
 def u32_to_i32(value):
