+++ alexk_low_cost_robot/requirements.txt
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:321756bd8afe9faedf8553c01128845e318a5a674a75f15bb89490577a3c4dd8
-size 41
+oid sha256:f18df0282dc36af7a027c5cdd68f32674ecf751f81e62b9d4618d04a06380123
+size 85
